⚡ 배송 송장번호 Reference Type 변경

See merge request s08-final/S08P31B306!169

+++ backend/reslow/src/main/java/com/ssafy/reslow/domain/order/dto/OrderUpdateCarrierRequest.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class OrderUpdateCarrierRequest {
 
-	private int carrierTrack;
+	private Long carrierTrack;
 	private String carrierCompany;
 }
